Share build metadata helper
Agent-Logs-Url: https://github.com/FredYakumo/zihuan-next/sessions/482cb624-6ac4-41bd-a43a-6132a43702ac
